requirements.txt
django
djangorestframework
pyyaml
requests
django-cors-headers
PyJWT
pytz
sqlparse
tzdata
asgiref
django-environ
djangorestframework-simplejwt
python-dotenv